Downgraded blob table from using double to using float (for parity)

+++ Game/Point.pde
@@ -1,7 +1,7 @@
 public class Point{
-  public double x, y;
+  public float x, y;
   
-  public Point(double x, double y){
+  public Point(float x, float y){
     this.x = x;
     this.y = y;
   }
